subir

+++ backend/controllers/tareasControllers.js
@@ -3,7 +3,7 @@ const Tarea = require('../models/tareasModels')
 
 const getTareas = asyncHandler(async (req, res) => {
 
-    const tareas = await Tarea.find({})
+    const tareas = await Tarea.find({ user: req.user.id })
 
     res.status(200).json(tareas)
 })
@@ -16,7 +16,8 @@ const createTareas = asyncHandler(async (req, res) => {
     }
 
     const tarea = await Tarea.create({
-        descripcion: req.body.descripcion
+        descripcion: req.body.descripcion,
+        user: req.user.id
     })
 
     res.status(201).json(tarea)
@@ -31,9 +32,14 @@ const updateTareas = asyncHandler(async (req, res) => {
         throw new Error('Esa tarea no existe')
     }
 
-    const tareaUpdated = await Tarea.findByIdAndUpdate(req.params.id, req.body, { new: true })
-
-    res.status(200).json(tareaUpdated)
+    //Nos aseguramos que solo el dueño  de la tarea la pueda modificar
+    if (tarea.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error("Usuario no Autorizado")
+    } else {
+        const tareaUpdated = await Tarea.findByIdAndUpdate(req.params.id, req.body, { new: true })
+        res.status(200).json(tareaUpdated)
+    }
 })
 
 const deleteTareas = asyncHandler(async (req, res) => {
@@ -45,10 +51,14 @@ const deleteTareas = asyncHandler(async (req, res) => {
         throw new Error('Esa tarea no existe')
     }
 
-    await Tarea.deleteOne(tarea)
-
-
-    res.status(200).json({ id: req.params.id })
+    //Nos aseguramos que solo el dueño  de la tarea la pueda eliminar
+    if (tarea.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error('Usuario no Autorizado')
+    } else {
+        await Tarea.deleteOne(tarea)
+        res.status(200).json({ id: req.params.id })
+    }
 })
 
 module.exports = {
@@ -57,6 +67,5 @@ module.exports = {
     updateTareas,
     deleteTareas
 }
-
